# full-stack-asset-transfer-guide/applications/rest-api/src/coconiko/coconiko_swagger.yaml
openapi: 3.0.1
info:
  title: Coconiko API
  version: 1.1.0
  description: |
    FabricネットワークにおけるユーザーとCoconikoのコインとNFTなどを管理するためのAPI。

    アプリからの呼び出しに最適化しており、ヘッダーに`X-API-Key`とログインしている`userId`を設定し、
    APIを呼び出す仕様となっています。
    - `X-Api-Key`で呼び出し元のシステムを識別しています
    - `userId`でログインしているユーザーを識別しています

    各APIの必要な認証情報は各APIの右側のロックアイコンで確認できます
    
    注意：「ユーザーID」と「アカウントID」が違います。
    - 「アカウントID」はFabricネットワーク上でユーザーを識別するIDです、新規ユーザーを登録する時に割り当てられます。`0x0`を保留しています。
    - 「ユーザーID」はアプリ側でユーザーを識別するIDです
...
paths:
...
  /api/coconiko/user:
    put:
      tags:
        - Coconiko
      summary: 新規ユーザーを登録
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - role
              properties:
                username:
                  type: string
                role:
                  type: string
                  description: ユーザーの役割
                  enum:
                    - user
                    - admin
              example:
                username: "test"
                role: "user"
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: 正常
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_name:
                    type: string
                    description: ユーザー名
                  app_id:
                    type: string
                    description: アプリケーションID
                  account:
                    type: object
                    description: アカウント情報
                    properties:
                      id:
                        type: string
                        description: アカウントID
                      role:
                        type: string
                        description: ユーザーの役割
                  timestamp:
                    type: string
                    format: date-time
                example:
                  user_name: "test"
                  app_id: "coconiko"
                  account:
                    id: "x509::/OU=client/CN=test.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA"
                    role: "user"
                  timestamp: "2022-01-01T00:00:00.000Z"
        '201':
          description: ユーザーが存在しています
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
    get:
      tags:
        - Coconiko
      summary: ユーザー情報を取得
      security:
        - ApiKeyAuth: []
          userId: []
      responses:
        '200':
          description: 正常
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    $ref: '#/components/schemas/UserInfo'
              example: {
                result: {
                  userId: "test",
                  accountid: "x509::/OU=client/CN=test.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA",
                  role: "user",
                  balance: 0,
                  activate: true
                }
              }
    post:
      tags:
        - Coconiko
      summary: ユーザー情報を更新
      security:
        - ApiKeyAuth: []
          userId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - active
              properties:
                active:
                  type: boolean
                  description: 有効
      responses:
        '200':
          description: 正常
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    $ref: '#/components/schemas/UserInfo'
              example: {
                result: {
                  userId: "test",
                  accountid: "x509::/OU=client/CN=test.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA",
                  role: "user",
                  balance: 0,
                  activate: true
                }
              }
...
  /api/coconiko/nft/Mint:
    post:
      tags:
        - Coconiko NFT
      summary: NFTを新規発行
      description: メタデータを指定して新しいNFTを発行
      security:
        - ApiKeyAuth: []
          userId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - metadata
              properties:
                metadata:
                  type: object
                  description: NFTメタデータ
                  properties:
                    name:
                      type: string
                      example: "限定版アート作品"
                    price:
                      type: number
                      example: 100
                    description: 
                      type: string
                      example: "デジタルアート作品の限定版"
                    image:
                      type: string
                      format: uri
                      example: "https://example.com/artwork.png"
      responses:
        '200':
          description: NFT発行成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                example:
                  "result": {
                    "docType": "coconiko-nft",
                    "id": "nft_2ta80u30t9xm82lgi77",
                    "owner": "x509::/OU=client/CN=test.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA",
                    "creator": "x509::/OU=client/CN=test.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA",
                    "metadata": {
                      "name": "限定版アート作品",
                      "price": 100,
                      "description": "デジタルアート作品の限定版",
                      "image": "https://example.com/artwork.png"
                    },
                    "created": "2025-03-10T05:00:43.171Z",
                    "lastUpdated": "2025-03-10T05:00:43.171Z",
                    "burned": false
                  }
  /api/coconiko/nft/Transfer:
    post:
      tags:
        - Coconiko NFT
      summary: NFTを転送
      description: 指定したNFTの所有権を転送
      security:
        - ApiKeyAuth: []
          userId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tokenId
                - to
              properties:
                tokenId:
                  type: string
                  description: NFTトークンID
                  example: "NFT_001"
                from:
                  type: string
                  description: 送信元アカウントID（オプション）、なければ登録しているユーザーのアカウントIDを利用します。
                  example: "x509::/OU=client/CN=user1.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA"
                to:
                  type: string
                  description: 送信先アカウントID
                  example: "x509::/OU=client/CN=user2.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA"
      responses:
        '200':
          description: 転送成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
  /api/coconiko/nft/{tokenId}:
    get:
      tags:
        - Coconiko NFT
      summary: NFT情報を取得
      description: 指定したNFTの詳細情報を取得
      security:
        - ApiKeyAuth: []
          userId: []
      parameters:
        - in: path
          name: tokenId
          schema:
            type: string
          required: true
          description: NFTトークンID
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    properties:
                      tokenId:
                        type: string
                      metadata:
                        type: object
                      currentOwner:
                        type: string
                      created:
                        type: string
                        format: date-time
                      transactionHistory:
                        type: array
                        items:
                          type: object
                          properties:
                            from:
                              type: string
                            to:
                              type: string
                            timestamp:
                              type: string
                              format: date-time
                example:
                  result:
                    tokenId: "NFT_001"
                    metadata:
                      name: "限定版アート作品"
                      price: 100
                      description: "デジタルアート作品の限定版"
                      image: "https://example.com/artwork.png"
                    currentOwner: "x509::/OU=client/CN=user2.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA"
                    created: "2024-01-01T00:00:00.000Z"
                    transactionHistory:
                      - from: "0x0"
                        to: "x509::/OU=client/CN=user1.coconiko..."
                        timestamp: "2024-01-01T00:00:00.000Z"
                      - from: "x509::/OU=client/CN=user1.coconiko..."
                        to: "x509::/OU=client/CN=user2.coconiko..."
                        timestamp: "2024-01-02T00:00:00.000Z"
  /api/coconiko/nft/my/NFTs:
    get:
      tags:
        - Coconiko NFT
      summary: ユーザーのNFT一覧情報を取得
      description: ユーザーのNFT一覧情報を取得
      security:
        - ApiKeyAuth: []
          userId: []
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: array
                    items:
                      type: object
                example:
                  result:
                    [
                      {
                        "nft_id": "nft_2ta80u30t9xm82lgi77",
                        "owner": "x509::/OU=client/CN=test.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA",
                        "creator": "x509::/OU=client/CN=test.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA",
                        "metadata": {
                          "name": "限定版アート作品",
                          "price": 100,
                          "description": "デジタルアート作品の限定版",
                          "image": "https://example.com/artwork.png"
                        },
                        "created_at": "2025-03-10T05:00:43.171Z",
                        "burned": false
                      },
                      {
                        "id": "nft_2ta80u30t9xm82lgi78",
                        "owner": "x509::/OU=client/CN=test.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA",
                        "creator": "x509::/OU=client/CN=test.coconiko::/C=JP/ST=Tokyo/O=Social Design Lab./OU=Local Coin Net/CN=Local Coin Net SDL Root CA",
                        "metadata": {
                          "name": "限定版アート作品",
                          "price": 200,
                          "description": "デジタルアート作品の限定版",
                          "image": "https://example.com/artwork.png"
                        },
                        "created_at": "2025-03-10T05:00:43.171Z",
                        "burned": false
                      }                      
                    ]
...
components:
  schemas:
...
    Result:
      type: object
      properties:
        result:
          type: object
          description: ページングされたクエリ結果
    UserInfo:
      type: object
      properties:
        userId:
          type: string
          description: ユーザーID
        accountId:
          type: string
          description: アカウントID
        role:
          type: string
          description: ユーザーの役割
        balance:
          type: integer
          description: ユーザーの残高
        active:
          type: boolean
          description: ユーザーの有効状態       
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: APPのAPIキー
    userId:
      type: apiKey
      in: header
      name: userId
      description: ログインしているユーザーID